refactor(mouse): move mouse cursor out of vidfade
