+++ src/renderer/IUIRenderer.cpp
@@ -22,7 +22,8 @@
 #include "bflib_vidraw.h"        // sprite and primitive CPU fallbacks
 #include "bflib_sprite.h"        // TbSprite
 #include "bflib_video.h"         // lbDisplay, units_per_pixel
-#include "vidmode.h"             // pixmap (TbColorTables) for fade_tables
+#include "vidmode.h"
+#include "vidfade.h"             // pixmap (TbColorTables) for fade_tables
 #include <algorithm>             // std::sort for the seq-merge replay
 #include <vector>
